Always read request body when using Rack
In certain versions of Passenger (all of them), leaving the request body
unread will cause an EPIPE exception inside Passenger, resulting in a
500 response. This is because Passenger makes a blocking write of the
request body to the application (Puppet). If we respond without reading
the body, this write is interrupted, which isn't handled properly by
Passenger. So when processing requests, we make sure to have always read
at least 1 character, which unblocks the write and lets Passenger
continue.

This issue is not present in other HTTP handlers.

https://code.google.com/p/phusion-passenger/issues/detail?id=471

Paired-With: Josh Cooper <josh@puppetlabs.com>