(#19558) Fix acceptance test failure
Without this patch applied the commit in 38ef0c5 causes a behavior
change in the way error messages are formatted that fail to pass through
the regular expression used in our Solaris 10 acceptance test suite.

This patch addresses the problem by updating the regular expression to
match the error message produced by the current behavior of the system.