Don't assume puppetbindir is defined
The host variable `puppetbindir` is undefined on FOSS windows (unlike
Unix), so this test was failing on Windows when trying to execute
`/ruby`. This is really a bug in the harness as the eventlog
acceptance test open-codes this logic as well.

This commit changes it to fallback to `ruby`. It also simplifies the
logic for detecting whether an agent supports SSLv2.