Don't require openssl client to return 0 on failure
Previously, the test assumed openssl s_client would return 0 when the
master rejected the SSLv2 connection, which is a bit odd.

This commit ignores the exit code when determining if the passed passed
or failed, and instead requires that a CLIENT-HELLO message be present,
and a SERVER-HELLO message is absent.