Don't assume master supports SSLv2
Previously, the test tried to create an SSLv2 rogue master, and then
check each agent's suitability (as to whether it's ruby supports SSLv2).
But this fails on precise, as its ruby does not support SSLv2.
Specifically the call to ssl_version = 'SSLv2' raises an exception.

Since there is nothing special about the master, we just need to run a
ruby process, this commit moves the rogue ssl server to the agent
itself, and only those agents where SSLv2 is available.

Since the agent may also be a master, it changes the rogue server's
accept port from 8140 to 8150.