(#13859) FreeBSD rc.conf variable are not set when a similar variable exists.
The regular expressions for finding the proper rc.conf _enable
variable to check, set, modify are too loose and will match on
similar variables causing unexpectd results.  For example, if
openprog_enable exists, you will not be able to enable the service
for prog.  This patch modifies the regular expressions for finding
the proper _enable variable to start the match at the beginning of
the line.