(#19716) Document handling of duplicate cron jobs
Prior to this commit the documentation claimed that if a user specified a
`cron` resources that exactly duplicated an existing unnamed entry in the
crontab then Puppet would attach a name to that entry and start managing it.

`cron` hasn't worked this way since at least 0.22.2.

Current behavior is: Puppet will remove all crontab entries duplicated by a `cron`
resource with `ensure => absent`. Puppet will preserve all entries duplicated
by a `cron` resource with `ensure => present` and treat the resource as a noop.

Mix the two cases and undefined behavior results.