(#11331) Add tracking of position on line
This contains a new feature; keeping track of the line position
to enable better reporting for syntax errors (e.g "line 13:3",
vs just "line 13"). This is helpful to a user since a particular
punctuation character may appear many times on a line.

The puppet/parser classes have been updated to at least accept
a pos. (The "old" lexer does not produce pos information, and if
pos is nil, it is ignored later).

The egrammar implementation is refactored, and instead of extracting
line information separately, it simply passes on an instance of
the SourcePosAdapter which has all the required information.

The egrammar lexer computes offsets, and positions. It stores them
in the lexing_context until asked for a detailed position.

The lexer needs to be multibyte aware when computing pos, and it is now
updated to use multibyte regular expressions for whitespace (as a side
effect). Multibyte is supported for Ruby 1.9.3 and later.