(#11331) Add eparser unit test.
Eparser was unit tested by the parser unit test, but did not work
as intended. This adds a separat test

This is basically a copy of the parser unit test, but with removal
of tests that only check the parser implementation (internal calls
to "self" from the old parser).

There are also minor adjustments as the egrammar parser has better
information (e.g. the result contains "file" instead of "File" as type
is already detected and type internally has already done a downcase).