(#11331) Add monkey-patch that fixes Rgen issue.
Rgen 0.6.1 contains an implementation that extends Array with
a method missing. In the 0.6.1 version of Rgen, this method fails
to honor the Ruby Hash protocol that the method "to_hash" on any
Object should produce a Hash. This problem only occured when
creating a Hash using Hash[[]] (i.e. to return an empty hash).

This addes two methods to array:
to_hash (which raises NoMethodError (i.e. does not change any behavior).
respond_to? which returns false for :to_hash.

This monkey patch should be removed when RGen releases a new version of
lib/rgen/array_extensions.rb that handles to_hash correctly (expected to
be in RGen 0.6.2).

The monkey patch is harmless since it basically adds an implementation
of what Array does if this patch was not there.