(#11331) Fix issues relating to running tests for Ruby 1.8.7.
There were several problems related to EvaluationError constant.
This is now simplified. Moving inclusion of module in tests to be
inside describe instead of done as rspec configuration.

Fix issues relating to Ruby 1.8.7 Hash not being in insert order.

Since Ruby Hash in 1.8.7 is not in insert order when iterating
several changes where required.
- dumping now sorts keys before dumping (this to get the same behavior
on all Ruby versions.
- checking evaluation result differs since result can be in any order -
test updated to check multiple permutations (this since it is not
meaningful to sort all evaluation results and they may be nested).

Fix char.upcase issues (not supported in Ruby 1.8.7).

This means using s[0,1].upcase instead of s[0].upcase to force it to be
a one char string. (Works in both versions).