(#2251) crontab: avoid duplication of cronjobs
The crontab provider will duplicate resources ad nauseam if another
record with the same "puppet name" exists in a different crontab.

It was found that entries in different crontabs, even when sharing
names, should be treated as independent entities. Therefor, the
crontab provider will now ensure that when prefetching crontabs
from disk, cronjobs are matched to managed resources only if both
the name and the target file match.