Correct formulation in AIX lsuser command
Currently, the command string composition for the lsuser command is incorrect.
It attempts to add a flag (-c) as part of the first binary to execute, when
puppet expects flags to separate elements in the command string array.. This
commit updates the forumlation so that "-c" is a separate element. How it ever
worked as it is now is a mystery.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>