(#3220) crontab: allow purging unmanaged resources
The problem was that unmanaged cronjobs would be parsed, but
not implicitly named.

Workaround: Pick an implicit name. It's derived from the command.
The provider will never put this implicit name into a crontab file
when writing it back to disk.

The names are not unique if several unmanaged jobs execute the same
command. That's a wrinkle and not mission critical, but it can
be surprising for users, because the "duplicates" will not be
purged during the first run.  If we want to go for this solution,
we should add a uniquifier.

TODO: wtf, when the integration test runs, it leaves parts
of the crontab fixture in ./<my-user-name>.
Not yet reproduced outside of rspec.