(#14764) Fix formatting of file mode audits
When a File resource is set to audit all attributes:

    file{"/tmp/foo": audit => "all"}

If the file moves between `ensure => present` and `ensure => absent` the value
of the `mode` parameter will be set to `:absent`.  This will cause the audit to
fail as the `is_to_s` method for file modes cannot comprehend symbols as it
expects a 3 or 4 digit string of numbers.

This commit causes `is_to_s` to delegate to `super` when `:absent` is passed.