Explicitly require Win32API
Previously, puppet could not be loaded in ruby 1.9 on Windows. The issue
is that commit cd7d52a7b caused Win32API to be called during puppet's
bootstrapping process, before Puppet.features.microsoft_windows? is
invoked, which normally handles loading Windows specific gems.

In ruby 1.8, this is not an issue because Win32API is always loaded by
ruby itself. But in ruby 1.9, it must be explicitly loaded, which this
commit does.