(#11331) Refactor computation of lexer line
This refactors the computation of line to make use of the line
index created at the start of lexing. 
This the entire input is scanned for lines there is no need to scan
for newlines in multiline comments, when handling the special RETURN
token, or when slurping strings.

This reduces the scanning overhead.

The ability for a token to set its incr_line attribute to true
is removed (it was only used by RETURN token) and this is now handled
generically.