(#656) keep crontab from re-applying certain commands
The cron type allows arbitrary text in its command property.
The crontab provider will silently discard leading and trailing spaces
in record fields. Therefor, when parsing the generated records from disk,
the spaces appear to be missing.

To avoid this, it is safe to munge the command and discard the spaces
on the type level. The provider never gets to see them and will not
assume the need to refresh.