(#11331) Fix issues with detailed errors for hostnames.
The validation of hostname expressions did not produce good messages
for interpolated string (syntax error vs. detailed error). Relaxes the
grammar to accept a concatenated string which is then validated.

Refactors parameter name from 'node' to 'semantic' in hostname_xxx
checks (the role of the parameter is to provide semantic location, it is
not required to be a node).

Contains additional tests for other (already passing) detailed errors.