(#15561) Fix spec failure in certificate_request_spec
Without this patch applied, the x.509 CN extraction improvements cause
the following spec failure in the certificate_request_spec.rb file

      1) Puppet::SSL::CertificateRequest when converting from a string should create a CSR ...
         Failure/Error: described_class.from_s("my csr")
         NoMethodError:
           undefined method `to_a' for "/CN=Foo.madstop.com":String
         # ./lib/puppet/util/ssl.rb:23:in `cn_from_subject'
         # ./lib/puppet/ssl/base.rb:54:in `name_from_subject'
         # ./lib/puppet/ssl/base.rb:62:in `from_instance'
         # ./lib/puppet/ssl/base.rb:71:in `from_s'
         # ./spec/unit/ssl/certificate_request_spec.rb:42:in `block (3 levels) in <top (required)>'

Poor stubbing is the root cause of this problem.  The mock certificate
object returns a String from `#subject`, however the system expects a
kind of `OpenSSL::X509::Name`.

This patch addresses the problem by changing the mock object to return
an instance of `OpenSSL::X509::Name` when the mocked `#subject` instance
is called.