(#15561) Make implementation match spec examples
Without this patch applied, the following spec failure occurs.

    3) Puppet::Util::SSL when getting a CN from a subject should return nil for a bare string
       Failure/Error: described_class.cn_from_subject("/CN=foo").should == nil
       NoMethodError:
         undefined method `to_a' for "/CN=foo":String
       # ./lib/puppet/util/ssl.rb:23:in `cn_from_subject'
       # ./spec/unit/util/ssl_spec.rb:47:in `block (3 levels) in <top (required)>'

This failure is caused by the implementation of the `cn_from_subject` method
calling `to_a` on the string, which raises an exception.

This patch makes the implementation match the specified behavior by checking
for the `to_a` method.  If false, `nil` is returned as expected.