(#12493) Add note that `user` defined in a scheduled_task resource must be a privileged user.
The scheduled_task resource includes a note about issues surrounding the use of the 'SYSTEM'
user and recommends choosing another user to run the scheduled task, or altering the security
policy to allow the 'SYSTEM' account to run scheduled tasks. The note does not mention
that puppet must be running as a privileged user to manage scheduled tasks in order to avoid
'access denied' errors.

This commit corrects the issue by adding a note that puppet must be running as a privileged
user to manage scheduled tasks and avoid 'access denied' errors.