(#1453) Document behavior of removed Cron params
Removing a parameter such as `minute => 10` from a manifest will not reset the
minute component of the associated cronjob to `*`. This is because Puppet only
manages parameters that are out of sync with manifest entries (see issue
[19198][19198]).

This patch updates the documentation for the Cron type to be explicit on this
point as it has been surprising to a lot of people.

  [19198]: https://projects.puppetlabs.com/issues/19198