Remove ServerName from Apache Virtualhost
I'm including this small patch because this mis-configuration caused
quite a bit of hurt for me.

Without this patch applied, both virtual hosts running on 8140 and 8141
use the same SSL certificate, the one configured in 8140.  This is a
problem because the second, 8141, virtual host is configured to use a
different certificate that is intended to be rejected by the agent.

Because the first cert is being used the agent does not know to reject
the connection.  This, in turn, caused a false-positive success in the
acceptance test that validates an agent refuses to connect to a rogue
master.

This patch fixes the problem by removing the ServerName statements
included in the virtualhost blocks.  With these statements removed from
the configuration the server on 8140 presents a cert issued by the
Master CA and the server on 8141 presents a cert issued by the Agent CA,
as expected.