(Maint) Don't use fails_on_windows to exclude spec tests
In the 2.7 series, we used the fails_on_windows rspec filter to
exclude unit tests on windows.

In 3.x, that approach has been removed. Instead, if the test should
never be expected to pass on Windows, then use an exclude filter like:

  describe Puppet::MyClass, :unless => Puppet.features.microsoft_windows? do

However, if the test doesn't currently pass, but should sometime in
the future, then use an rspec conditional pending block:

  pending("porting to Windows", :if => Puppet.features.microsoft_windows?) do
    <example1>
  end