Revert "Merge branch 'pull-1540'"
This is causing spec failures on windows, and the current implementation
of the crontab provider and the parsedfile provider is sufficiently
complex that it's not clear how to sanely fix this issue. This is
getting reverted until it's in a more understandable state.

This reverts commit f85ec37456095c06a7b0d99a752534d0caf9c97e, reversing
changes made to 138fc337c36715a7ab40b2b68b86aa71ec3ada64.